fix: small changes

+++ database/migrations/2025_07_14_073655_create_taxes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('taxes', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->double('rate');
+            $table->decimal('rate', 5);
             $table->timestamps();
         });
     }
